# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/articlecomment/{id}": {
            "delete": {
                "tags": [
                    "ArticleComment"
                ],
                "description": "Delete ArticleComment",
                "operationId": "articlecommentDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "commentid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
...
        "/api/article/{id}": {
            "delete": {
                "tags": [
                    "Article"
                ],
                "description": "Delete Article",
                "operationId": "articleDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
...
        "/api/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "description": "Index of Comment",
                "operationId": "commentIndex",
                "responses": {
                    "default": {
                        "description": "Success: Array of comments",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CommentData"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "comment"
                ],
                "description": "Store comment",
                "operationId": "comment",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "body",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "visible",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentStore"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentData"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/comment/{id}": {
            "put": {
                "tags": [
                    "Comment"
                ],
                "description": "Comment update",
                "operationId": "commentUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentUpdateResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comment"
                ],
                "description": "Delete Comments",
                "operationId": "commentDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CommentData": {
                "properties": {
                    "user_id": {
                        "description": "user_id",
                        "type": "integer"
                    },
                    "article_id": {
                        "description": "article_id",
                        "type": "integer"
                    },
                    "product_id": {
                        "description": "product_id",
                        "type": "integer"
                    },
                    "body": {
                        "description": "body",
                        "type": "string"
                    },
                    "visible": {
                        "description": "visible",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "CommentStore": {},
            "CommentUpdateResponse": {
                "properties": {
                    "product_id": {
                        "description": "product_id",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "user_id",
                        "type": "integer"
                    },
                    "article_id": {
                        "description": "article_id",
                        "type": "integer"
                    },
                    "body": {
                        "description": "body",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
...
    }
}